chore(e2e/utils): extract getDetoxConfig to utils
will need later for REPL
