+++ e2e/init.js
@@ -1,10 +1,7 @@
 const detox = require('detox')
-const detoxConfig = require('../package.json').detox
+const getDetoxConfig = require('./utils').getDetoxConfig
 const adapter = require('detox/runners/jest/adapter')
 const specReporter = require('detox/runners/jest/specReporter')
-
-const ADB = require('detox/src/devices/android/ADB')
-const adb = new ADB()
 
 // Set the default timeout
 jest.setTimeout(120000)
@@ -15,31 +12,8 @@ jasmine.getEnv().addReporter(adapter)
 jasmine.getEnv().addReporter(specReporter)
 
 beforeAll(async () => {
-  const configs = detoxConfig.configurations
-
-  // TODO figure out iOS configs
-  const newConfigs = detoxConfig.configurations = {
-    'ios.sim.debug': configs['ios.sim.debug']
-  }
-
-  // detox device configurations are generated dynamically here after querying
-  // ADB for android devices and emulator, instead of hardcoding in package.json
-
   try {
-    const devices = await adb.devices()
-    devices.forEach(device => {
-      const key = 'android' + (device.type == 'emulator' ? '.emu' : '')
-      const releaseTypes = ['debug', 'release']
-      releaseTypes.forEach(releaseType => {
-        const configKey = `${key}.${releaseType}`
-        newConfigs[configKey] = configs[configKey]
-        newConfigs[configKey].name = device.name
-      })
-    })
-  } catch(err) {
-    console.error("Could not find android device/emulator", err)
-  }
-  try {
+    const detoxConfig = await getDetoxConfig()
     await detox.init(detoxConfig)
   } catch (err) {
     // when detox init fails we should really stop the tests otherwise we get a
